fix(status): slim ambient appearance bundle
